requirements.txt
python-telegram-bot==13.*
sqlalchemy==1.4.*
sqlalchemy-firebird==0.7.*